feat(backend): Provide wallet identity endpoint
Description

The wallet public key, address and connected node should be available to the user.
If docker wallet container is running we may call grpc endpoint (identify) in order to get:
    public_key
    public_address
    node_id

Motivation and Context

User should be allowed to use his wallet or to share it.

How Has This Been Tested?

Manually:
Started the Launchpad app.
Tor, Base Node and Wallet are started.
Wallet identity is printed.
Wallet container is stopped and error message is printed.