feat(backend): run wallet grpc server with a password. Description
We will run the grpc server using a session password: the password is provided when the wallet is started.
Or the password could be configred using the conf. variable: TARI_WALLET_GRPC_PASSWORD.

Motivation and Context
We should protect the wallet sensitive infromation.

How Has This Been Tested?
Both cases are maunally tested.
Run the wallet with -- --grpc-password=12345678A and called getBalance.
When the correct password is provide balance is shown.
If not then 16 UNAUTHENTICATED: Invalid gRpc password is shown.

Run without a password and balance is shown with or without authorization.