Add new image definitions for grafana
We're replacing fronttail with grafana.

This commit adds the ImageType definitions for Loki, Promtail and
grafana.

The configuration files for grafana are included in the launchpad assets
folder. These will be mounted so that the containers can access them,
but will be in the host file system so that advanced users can tweak
things.

I've also tidied up the `mount` definitions to allow arbitrary mount
bindings, which grafana is going to need (coming soon).

The front-end should be updated to derive the image list using the
`image_list` command rather than hard-coding the list, so that these
changes can be picked up. The FE still reports fronttail.