feat: use grafana as logging backend
We're replacing fronttail with grafana.

This PR adds the ImageType definitions for Loki, Promtail and
grafana and removes all traces of frontail. Thank you for your service,
but you're fired.

The configuration files for grafana are included in the launchpad assets
folder. These will be mounted so that the containers can access them,
but will be in the host file system so that advanced users can tweak
things.

The front-end should be updated to derive the image list using the
`image_list` command rather than hard-coding the list, so that these
changes can be picked up.
I've updated the hard-coded image list, but it's not the most elegant
solution.

I've also added settings to get grafana containers working
* Touching up defaults.ini, configuring the correct ports and netwroking
  and disabling analytics
* fixing the PATH envar for grafana (doh!)
* adding a loki config file