feat(backend): show wallet emoji id and start streaming when wallet is started Description
feature: #257
Show both UUID and emoji representation.

feature: #297
Subscribe for events and wait for wallet.

Motivation and Context
Waiting for the wallet on the backend should be temproray solution.
Emoji id is a core Tari feature.

How Has This Been Tested?
Started the launchpad.
There are not printed edenity info or wallet transactions in the console.
Start the wallet.
Idenity is printed along with emoji id.
Streaming is started.