feat: pull docker images reporting
Description
Feature: #311

We should provide meta information about the docker images.
If images are not up to date(latest) then we should pull and start the latest ones.

Motivation and Context

User should be able to upgrade to the latest image or to start a custom image.

How Has This Been Tested?
This is not tested as it is in progress.