feat: sync blockchain
Description:

Listen to the event from BaseNode on the last step of the onboarding.
Providing following information to the user:

	progress (0-100%)
	the time estimation

Motivation and Context

How Has This Been Tested?
Manually.
Test stream listener is created.
When base node is started it the listener connects and read progress info calling grpc method: get_sync_progress.
Events are printed in the console.